Remove socket from rpc test and move integration test
